Removed unnecessary imports from app.py. Renamed functions.py color_functions.py for clarity.
