+++ lib/color_functions.py
